refactor: code audit optimizations (5 fixes)
1. sanitize_filename dedup → utils/sanitize.py (was duplicated in
   cli.py and web/utils.py)
2. SOFT_STATUS_STEPS ↔ STATUS_FIELD_FOR_STEP assert (catches typos
   at import time)
3. SentenceTransformer lru_cache (was reloaded twice per match_clips
   call, saves 1-3 seconds)
4. SRT UTF-8 BOM (utf-8-sig) — prevents CJK mojibake on older players
5. resolve_video elif + comments (makes fallthrough explicit)

Audit items skipped (with reasoning):
- #2 _ALLOWED_TOP: manual check provides better error messages
- #6 TTS LRU O(n log n): premature optimization, unnoticeable
- #7 CLI top-level imports: low impact, high effort
- #8 retry loop: no bug (audit confirms)
- #9 ScriptSegment single field: YAGNI
